+++ CHANGELOG.md
@@ -18,6 +18,20 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 - Python 3.9 support.
 
+- `sdk.cases` methods:
+    - `sdk.cases.create()`
+    - `sdk.cases.get_all()`
+    - `sdk.cases.get_page()`
+    - `sdk.cases.export_summary()`
+    - `sdk.cases.get()`
+    - `sdk.case.update()`
+
+- `sdk.cases.file_events` methods for managing file events associated with a given case:
+    - `sdk.cases.file_events.add()`
+    - `sdk.cases.file_events.delete()`
+    - `sdk.cases.file_events.get()`
+    - `sdk.cases.file_events.get_all()`
+
 ## 1.10.1 - 2020-12-16
 
 ### Fixed